+++ lib/mix/tasks/cinder.upgrade.ex
@@ -1,6 +1,8 @@
 if Code.ensure_loaded?(Igniter) do
   defmodule Mix.Tasks.Cinder.Upgrade do
-    @moduledoc false
+    @moduledoc """
+    Tasks for automatic migration of your code between various versions of Cinder.
+    """
 
     use Igniter.Mix.Task
 
@@ -33,6 +35,7 @@ if Code.ensure_loaded?(Igniter) do
 
       upgrades =
         %{
+          "0.9.0" => [&flatten_theme_component_blocks/2],
           "0.10.0" => [&rename_boolean_theme_keys/2]
         }
 
@@ -40,6 +43,123 @@ if Code.ensure_loaded?(Igniter) do
         custom_opts: options
       )
     end
+
+    # --- 0.9.0: Flatten component/2 blocks to flat set calls ---
+
+    @doc false
+    def flatten_theme_component_blocks(igniter, _opts) do
+      Igniter.update_all_elixir_files(igniter, fn zipper ->
+        with {:ok, _zipper} <- Igniter.Code.Module.move_to_module_using(zipper, [Cinder.Theme]) do
+          # Work on the source text directly since this is a structural transform
+          source_string = zipper |> Sourceror.Zipper.topmost_root() |> Sourceror.to_string()
+          new_string = transform_component_blocks(source_string)
+
+          if new_string == source_string do
+            {:ok, zipper}
+          else
+            case Sourceror.parse_string(new_string) do
+              {:ok, new_ast} ->
+                new_zipper = Sourceror.Zipper.zip(new_ast)
+                {:ok, new_zipper}
+
+              {:error, _} ->
+                {:ok, zipper}
+            end
+          end
+        else
+          _ -> {:ok, zipper}
+        end
+      end)
+    end
+
+    @doc false
+    def transform_component_blocks(source_string) do
+      case Sourceror.parse_string(source_string) do
+        {:ok, ast} ->
+          patches = collect_component_patches(ast)
+
+          if patches == [] do
+            source_string
+          else
+            source_string
+            |> Sourceror.patch_string(patches)
+            |> Code.format_string!(locals_without_parens: [set: 2, extends: 1])
+            |> IO.iodata_to_binary()
+          end
+
+        {:error, _} ->
+          source_string
+      end
+    end
+
+    defp collect_component_patches(ast) do
+      {_ast, patches} =
+        Macro.prewalk(ast, [], fn
+          # Pattern for Sourceror AST: do block is stored as {{:__block__, _, [:do]}, body}
+          {:component, _meta, [module_ast, [{{:__block__, _, [:do]}, body}]]} = node, acc ->
+            extract_and_patch_component(node, module_ast, body, acc)
+
+          # Pattern for standard AST: do block is stored as {:do, body}
+          {:component, _meta, [module_ast, [do: body]]} = node, acc ->
+            extract_and_patch_component(node, module_ast, body, acc)
+
+          node, acc ->
+            {node, acc}
+        end)
+
+      Enum.reverse(patches)
+    end
+
+    defp extract_and_patch_component(node, module_ast, body, acc) do
+      comment = module_to_comment(module_ast)
+
+      set_calls =
+        case body do
+          {:__block__, _, calls} -> calls
+          single_call -> [single_call]
+        end
+
+      replacement_text = build_replacement_text(comment, set_calls)
+      range = Sourceror.get_range(node)
+
+      if range do
+        patch = %{range: range, change: replacement_text}
+        {node, [patch | acc]}
+      else
+        {node, acc}
+      end
+    end
+
+    @doc false
+    def module_to_comment(module_ast) do
+      case module_ast do
+        {:__aliases__, _, parts} ->
+          parts |> List.last() |> to_string()
+
+        atom when is_atom(atom) ->
+          atom |> Module.split() |> List.last()
+
+        _ ->
+          "Component"
+      end
+    end
+
+    defp build_replacement_text(comment, set_calls) do
+      set_call_strings =
+        Enum.map(set_calls, fn call ->
+          Sourceror.to_string(call)
+        end)
+
+      case set_call_strings do
+        [] ->
+          "# #{comment}"
+
+        calls ->
+          "# #{comment}\n" <> Enum.join(calls, "\n")
+      end
+    end
+
+    # --- 0.10.0: Rename filter_boolean_* theme keys to filter_radio_group_* ---
 
     @doc false
     def rename_boolean_theme_keys(igniter, _opts) do
@@ -68,7 +188,6 @@ if Code.ensure_loaded?(Igniter) do
         )
       end)
       |> Enum.reduce(zipper, fn _found_zipper, acc_zipper ->
-        # Re-find the node from the current accumulator zipper
         case Igniter.Code.Common.move_to(acc_zipper, fn z ->
                match?(
                  {:set, _, [{:__block__, _, [^old_key]}, _]},